Move checkpoint race test into checkpoint_test.go
Address review feedback: consolidate into the package's canonical test file.
